ZeonCharity/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.1
django-filter==22.1
django-heroku==0.3.1
django-jazzmin==2.5.0
django-nested-admin==3.4.0
django-nested-inline==0.4.5
django-phonenumber-field==6.4.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.54
phonenumberslite==8.12.54
Pillow==9.2.0
psycopg2==2.9.3
pyparsing==3.0.9
python-decouple==3.6
python-monkey-business==1.0.0
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
pandas==1.5.0
numpy==1.23.3
et-xmlfile==1.1.0 
openpyxl==3.0.10
